requirements.txt
setuptools
datetime
ipaddress
requests
paramiko
six
gitpython
urllib3
uuid
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
behave
behave2cucumber
argparse
jsonschema
PyHamcrest
beautifulsoup4
behave-py3
pyyaml
suds-py3
xmltodict
deepdiff
python-gitlab
flask_debug
cffi
pysnmp
jira
